+++ pkg/query/querier_test.go
@@ -19,6 +19,7 @@ import (
 	"github.com/fortytw2/leaktest"
 	"github.com/go-kit/kit/log"
 	"github.com/pkg/errors"
+	"github.com/prometheus/prometheus/pkg/gate"
 	"github.com/prometheus/prometheus/pkg/labels"
 	"github.com/prometheus/prometheus/pkg/timestamp"
 	"github.com/prometheus/prometheus/pkg/value"
@@ -38,16 +39,16 @@ type sample struct {
 }
 
 func TestQueryableCreator_MaxResolution(t *testing.T) {
-	defer leaktest.CheckTimeout(t, 10*time.Second)()
+	t.Cleanup(leaktest.CheckTimeout(t, 10*time.Second))
 	testProxy := &storeServer{resps: []*storepb.SeriesResponse{}}
-	queryableCreator := NewQueryableCreator(nil, testProxy)
+	queryableCreator := NewQueryableCreator(nil, nil, testProxy, 2, 5*time.Second)
 
 	oneHourMillis := int64(1*time.Hour) / int64(time.Millisecond)
 	queryable := queryableCreator(false, nil, oneHourMillis, false, false)
 
 	q, err := queryable.Querier(context.Background(), 0, 42)
 	testutil.Ok(t, err)
-	defer func() { testutil.Ok(t, q.Close()) }()
+	t.Cleanup(func() { testutil.Ok(t, q.Close()) })
 
 	querierActual, ok := q.(*querier)
 
@@ -58,7 +59,7 @@ func TestQueryableCreator_MaxResolution(t *testing.T) {
 
 // Tests E2E how PromQL works with downsampled data.
 func TestQuerier_DownsampledData(t *testing.T) {
-	defer leaktest.CheckTimeout(t, 10*time.Second)()
+	t.Cleanup(leaktest.CheckTimeout(t, 10*time.Second))
 	testProxy := &storeServer{
 		resps: []*storepb.SeriesResponse{
 			storeSeriesResponse(t, labels.FromStrings("__name__", "a", "zzz", "a", "aaa", "bbb"), []sample{{99, 1}, {199, 5}}),                   // Downsampled chunk from Store.
@@ -70,12 +71,12 @@ func TestQuerier_DownsampledData(t *testing.T) {
 		},
 	}
 
-	q := NewQueryableCreator(nil, testProxy)(false, nil, 9999999, false, false)
-
+	timeout := 10 * time.Second
+	q := NewQueryableCreator(nil, nil, testProxy, 2, timeout)(false, nil, 9999999, false, false)
 	engine := promql.NewEngine(
 		promql.EngineOpts{
 			MaxSamples: math.MaxInt32,
-			Timeout:    10 * time.Second,
+			Timeout:    timeout,
 		},
 	)
 
@@ -493,9 +494,10 @@ func TestQuerier_Select(t *testing.T) {
 			},
 		},
 	} {
+		timeout := 5 * time.Second
 		e := promql.NewEngine(promql.EngineOpts{
 			Logger:     logger,
-			Timeout:    5 * time.Second,
+			Timeout:    timeout,
 			MaxSamples: math.MaxInt64,
 		})
 
@@ -507,16 +509,15 @@ func TestQuerier_Select(t *testing.T) {
 				{dedup: false, expected: tcase.expected},
 				{dedup: true, expected: []series{tcase.expectedAfterDedup}},
 			} {
-				q := newQuerier(context.Background(), nil, tcase.mint, tcase.maxt, tcase.replicaLabels, tcase.storeAPI, sc.dedup, 0, true, false)
-				defer testutil.Ok(t, q.Close())
+				g := gate.New(2)
+				q := newQuerier(context.Background(), nil, nil, tcase.mint, tcase.maxt, tcase.replicaLabels, tcase.storeAPI, sc.dedup, 0, true, false, g, timeout)
+				t.Cleanup(func() { testutil.Ok(t, q.Close()) })
 
 				t.Run(fmt.Sprintf("dedup=%v", sc.dedup), func(t *testing.T) {
 					t.Run("querier.Select", func(t *testing.T) {
-						defer leaktest.CheckTimeout(t, 10*time.Second)()
-						defer testutil.Ok(t, q.Close())
+						t.Cleanup(leaktest.CheckTimeout(t, 10*time.Second))
 
 						res := q.Select(false, tcase.hints, tcase.matchers...)
-
 						testSelectResponse(t, sc.expected, res)
 
 						if tcase.expectedWarning != "" {
@@ -527,12 +528,12 @@ func TestQuerier_Select(t *testing.T) {
 					})
 					// Integration test: Make sure the PromQL would select exactly the same.
 					t.Run("through PromQL with 100s step", func(t *testing.T) {
-						defer leaktest.CheckTimeout(t, 10*time.Second)()
+						t.Cleanup(leaktest.CheckTimeout(t, 10*time.Second))
 
 						catcher := &querierResponseCatcher{t: t, Querier: q}
 						q, err := e.NewRangeQuery(&mockedQueryable{catcher}, tcase.equivalentQuery, timestamp.Time(tcase.mint), timestamp.Time(tcase.maxt), 100*time.Second)
 						testutil.Ok(t, err)
-						defer q.Close()
+						t.Cleanup(q.Close)
 
 						r := q.Exec(context.Background())
 						testutil.Ok(t, r.Err)
@@ -676,18 +677,24 @@ func TestQuerierWithDedupUnderstoodByPromQL_Rate(t *testing.T) {
 			"gprd", "fqdn", "web-08-sv-gprd.c.gitlab-production.internal", "instance", "web-08-sv-gprd.c.gitlab-production.internal:8083", "job", "gitlab-rails", "monitor", "app", "provider",
 			"gcp", "region", "us-east", "replica", "02", "shard", "default", "stage", "main", "tier", "sv", "type", "web",
 		)
-		q := newQuerier(context.Background(), logger, realSeriesWithStaleMarkerMint, realSeriesWithStaleMarkerMaxt, []string{"replica"}, s, false, 0, true, false)
-		defer func() { testutil.Ok(t, q.Close()) }()
+
+		timeout := 100 * time.Second
+		g := gate.New(2)
+		q := newQuerier(context.Background(), logger, nil, realSeriesWithStaleMarkerMint, realSeriesWithStaleMarkerMaxt, []string{"replica"}, s, false, 0, true, false, g, timeout)
+		t.Cleanup(func() {
+			testutil.Ok(t, q.Close())
+		})
 
 		e := promql.NewEngine(promql.EngineOpts{
 			Logger:     logger,
-			Timeout:    5 * time.Second,
+			Timeout:    timeout,
 			MaxSamples: math.MaxInt64,
 		})
 		t.Run("Rate=5mStep=100s", func(t *testing.T) {
+			t.Cleanup(leaktest.CheckTimeout(t, 10*time.Second))
+
 			q, err := e.NewRangeQuery(&mockedQueryable{q}, `rate(gitlab_transaction_cache_read_hit_count_total[5m])`, timestamp.Time(realSeriesWithStaleMarkerMint).Add(5*time.Minute), timestamp.Time(realSeriesWithStaleMarkerMaxt), 100*time.Second)
 			testutil.Ok(t, err)
-			defer q.Close()
 
 			r := q.Exec(context.Background())
 			testutil.Ok(t, r.Err)
@@ -715,9 +722,10 @@ func TestQuerierWithDedupUnderstoodByPromQL_Rate(t *testing.T) {
 			}, vec)
 		})
 		t.Run("Rate=30mStep=500s", func(t *testing.T) {
+			t.Cleanup(leaktest.CheckTimeout(t, 10*time.Second))
+
 			q, err := e.NewRangeQuery(&mockedQueryable{q}, `rate(gitlab_transaction_cache_read_hit_count_total[30m])`, timestamp.Time(realSeriesWithStaleMarkerMint).Add(30*time.Minute), timestamp.Time(realSeriesWithStaleMarkerMaxt), 500*time.Second)
 			testutil.Ok(t, err)
-			defer q.Close()
 
 			r := q.Exec(context.Background())
 			testutil.Ok(t, r.Err)
@@ -743,18 +751,24 @@ func TestQuerierWithDedupUnderstoodByPromQL_Rate(t *testing.T) {
 			"gprd", "fqdn", "web-08-sv-gprd.c.gitlab-production.internal", "instance", "web-08-sv-gprd.c.gitlab-production.internal:8083", "job", "gitlab-rails", "monitor", "app", "provider",
 			"gcp", "region", "us-east", "shard", "default", "stage", "main", "tier", "sv", "type", "web",
 		)
-		q := newQuerier(context.Background(), logger, realSeriesWithStaleMarkerMint, realSeriesWithStaleMarkerMaxt, []string{"replica"}, s, true, 0, true, false)
-		defer func() { testutil.Ok(t, q.Close()) }()
+
+		timeout := 5 * time.Second
+		g := gate.New(2)
+		q := newQuerier(context.Background(), logger, nil, realSeriesWithStaleMarkerMint, realSeriesWithStaleMarkerMaxt, []string{"replica"}, s, true, 0, true, false, g, timeout)
+		t.Cleanup(func() {
+			testutil.Ok(t, q.Close())
+		})
 
 		e := promql.NewEngine(promql.EngineOpts{
 			Logger:     logger,
-			Timeout:    5 * time.Second,
+			Timeout:    timeout,
 			MaxSamples: math.MaxInt64,
 		})
 		t.Run("Rate=5mStep=100s", func(t *testing.T) {
+			t.Cleanup(leaktest.CheckTimeout(t, 10*time.Second))
+
 			q, err := e.NewRangeQuery(&mockedQueryable{q}, `rate(gitlab_transaction_cache_read_hit_count_total[5m])`, timestamp.Time(realSeriesWithStaleMarkerMint).Add(5*time.Minute), timestamp.Time(realSeriesWithStaleMarkerMaxt), 100*time.Second)
 			testutil.Ok(t, err)
-			defer q.Close()
 
 			r := q.Exec(context.Background())
 			testutil.Ok(t, r.Err)
@@ -777,9 +791,10 @@ func TestQuerierWithDedupUnderstoodByPromQL_Rate(t *testing.T) {
 			}, vec)
 		})
 		t.Run("Rate=30mStep=500s", func(t *testing.T) {
+			t.Cleanup(leaktest.CheckTimeout(t, 10*time.Second))
+
 			q, err := e.NewRangeQuery(&mockedQueryable{q}, `rate(gitlab_transaction_cache_read_hit_count_total[30m])`, timestamp.Time(realSeriesWithStaleMarkerMint).Add(30*time.Minute), timestamp.Time(realSeriesWithStaleMarkerMaxt), 500*time.Second)
 			testutil.Ok(t, err)
-			defer q.Close()
 
 			r := q.Exec(context.Background())
 			testutil.Ok(t, r.Err)
@@ -800,7 +815,7 @@ func TestQuerierWithDedupUnderstoodByPromQL_Rate(t *testing.T) {
 }
 
 func TestSortReplicaLabel(t *testing.T) {
-	defer leaktest.CheckTimeout(t, 10*time.Second)()
+	t.Cleanup(leaktest.CheckTimeout(t, 10*time.Second))
 
 	tests := []struct {
 		input       []storepb.Series
@@ -867,7 +882,7 @@ func expandSeries(t testing.TB, it chunkenc.Iterator) (res []sample) {
 }
 
 func TestDedupSeriesSet(t *testing.T) {
-	defer leaktest.CheckTimeout(t, 10*time.Second)()
+	t.Cleanup(leaktest.CheckTimeout(t, 10*time.Second))
 
 	tests := []struct {
 		input       []series
@@ -1198,7 +1213,7 @@ func TestDedupSeriesSet(t *testing.T) {
 }
 
 func TestDedupSeriesIterator(t *testing.T) {
-	defer leaktest.CheckTimeout(t, 10*time.Second)()
+	t.Cleanup(leaktest.CheckTimeout(t, 10*time.Second))
 
 	// The deltas between timestamps should be at least 10000 to not be affected
 	// by the initial penalty of 5000, that will cause the second iterator to seek
@@ -1300,7 +1315,7 @@ type storeServer struct {
 	resps []*storepb.SeriesResponse
 }
 
-func (s *storeServer) Series(r *storepb.SeriesRequest, srv storepb.Store_SeriesServer) error {
+func (s *storeServer) Series(_ *storepb.SeriesRequest, srv storepb.Store_SeriesServer) error {
 	for _, resp := range s.resps {
 		err := srv.Send(resp)
 		if err != nil {